fix(core-library): do not create pr when a pr already exists for a ref (#66)
Recall that GitHub PRs just point to a reference head which are applied on-top of some base reference head, typically master, and so when the reference gets updated the PR automatically gets updated as well.

There may be scenarios where a pull request is already created for a ref, and a user will re-invoke code-suggester create pr using the same ref. If the `force` flag/parameter is true, a new ref is not created. Instead code-suggester create pr will force-update the existing reference. If we try to re-create the a pr for the same reference, there will be a duplication issue. Hence, if there is an existing pr, just update the ref, and GitHub will just adjust UI display accordingly to the reference HEAD data. However, if `force` flag/parameter is false, code-suggester create pr will fail when trying to create the reference because there is a duplicate reference.

**TL;DR**
When there is an existing pr for the a reference and code-suggester create pr is re-invoked, do not create a new pr. Instead, re-use the old one and update the ref.

Closes #17