feat(action): build as an javascript action (#331)
This switches the GitHub action to be a javascript action (from Docker). This is orders of magnitude faster to run -- no more docker build (or download) on every invocation.

The downside is that we must compile the dist file before each release. This is accomplished via a new workflow that runs on pushes to main that will open a new PR if there are changes to the dist output.

Fixes #141